feat: disable controls when dragging
- additionally splits up control panel into separate components
